pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stt-keyboard"
version = "1.0.0"
description = "System-wide offline speech-to-text keyboard input"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "STT Keyboard Team"}
]
keywords = ["speech-to-text", "voice-recognition", "keyboard", "dictation", "offline"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "PyQt6>=6.6.0",
    "vosk>=0.3.45",
    "sounddevice>=0.4.6",
    "keyboard>=0.13.5",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
stt-keyboard = "stt_keyboard.app:main"

[project.urls]
Homepage = "https://github.com/yourusername/stt-keyboard"
Repository = "https://github.com/yourusername/stt-keyboard"
Issues = "https://github.com/yourusername/stt-keyboard/issues"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
